Extrack useSetupMedia
