+++ src/create.tsx
@@ -1,45 +1,20 @@
-import type { JSX, ReactNode } from "react";
-import { useEffect, createContext, useCallback, useContext, useMemo, useRef, useSyncExternalStore } from "react";
+import type { JSX } from "react";
+import { createContext, useCallback, useContext, useMemo, useRef, useSyncExternalStore } from "react";
 import type { Selector, State, UseReconnection } from "./state.types";
-import type { PeerStatus, TrackId, TrackWithOrigin } from "./state.types";
 import type { ConnectConfig, CreateConfig } from "@fishjam-dev/ts-client";
 import type {
   DeviceManagerConfig,
-  Devices,
   CameraAPI,
   MicrophoneAPI,
   ScreenShareAPI,
-  UseSetupMediaConfig,
-  UseSetupMediaResult,
+  CreateFishjamClient,
+  FishjamContextType,
+  FishjamContextProviderProps,
+  UseConnect,
 } from "./types";
-import type { ClientApi, ClientEvents } from "./Client";
 import { Client } from "./Client";
-import type { MediaDeviceType, ScreenShareManagerConfig, TrackType } from "./ScreenShareManager";
-
-export type FishjamContextProviderProps = {
-  children: ReactNode;
-};
-
-type FishjamContextType<PeerMetadata, TrackMetadata> = {
-  state: State<PeerMetadata, TrackMetadata>;
-};
-
-export type UseConnect<PeerMetadata> = (config: ConnectConfig<PeerMetadata>) => () => void;
-
-export type CreateFishjamClient<PeerMetadata, TrackMetadata> = {
-  FishjamContextProvider: ({ children }: FishjamContextProviderProps) => JSX.Element;
-  useConnect: () => (config: ConnectConfig<PeerMetadata>) => () => void;
-  useDisconnect: () => () => void;
-  useStatus: () => PeerStatus;
-  useSelector: <Result>(selector: Selector<PeerMetadata, TrackMetadata, Result>) => Result;
-  useTracks: () => Record<TrackId, TrackWithOrigin<PeerMetadata, TrackMetadata>>;
-  useSetupMedia: (config: UseSetupMediaConfig<TrackMetadata>) => UseSetupMediaResult;
-  useCamera: () => Devices<TrackMetadata>["camera"];
-  useMicrophone: () => Devices<TrackMetadata>["microphone"];
-  useScreenShare: () => ScreenShareAPI<TrackMetadata>;
-  useClient: () => Client<PeerMetadata, TrackMetadata>;
-  useReconnection: () => UseReconnection;
-};
+import type { ScreenShareManagerConfig } from "./ScreenShareManager";
+import { createUseSetupMediaHook } from "./useSetupMedia";
 
 /**
  * Create a client that can be used with a context.
@@ -268,345 +243,6 @@ export const create = <PeerMetadata, TrackMetadata>(
     return state.devices.microphone;
   };
 
-  const useSetupMedia = (config: UseSetupMediaConfig<TrackMetadata>): UseSetupMediaResult => {
-    const { state } = useFishjamContext();
-    const configRef = useRef(config);
-
-    useEffect(() => {
-      configRef.current = config;
-
-      if (config.screenShare.streamConfig) {
-        state.client.setScreenManagerConfig(config.screenShare.streamConfig);
-      }
-
-      state.client.setDeviceManagerConfig({
-        storage: config.storage,
-      });
-    }, [config, state.client]);
-
-    useEffect(() => {
-      if (configRef.current.startOnMount && state.deviceManager.getStatus() === "uninitialized") {
-        state.devices.init({
-          audioTrackConstraints: configRef.current?.microphone?.trackConstraints,
-          videoTrackConstraints: configRef.current?.camera?.trackConstraints,
-        });
-      }
-      // eslint-disable-next-line
-    }, []);
-
-    const isBroadcastedDeviceChanged = (
-      expectedMediaDeviceType: MediaDeviceType,
-      client: ClientApi<PeerMetadata, TrackMetadata>,
-      pending: boolean,
-      mediaDeviceType: MediaDeviceType,
-    ) =>
-      client.status === "joined" && mediaDeviceType === expectedMediaDeviceType && !pending && !client.isReconnecting();
-
-    useEffect(() => {
-      let pending = false;
-
-      const broadcastOnCameraStart = async (
-        event: { mediaDeviceType: MediaDeviceType },
-        client: ClientApi<PeerMetadata, TrackMetadata>,
-      ) => {
-        const config = configRef.current.camera;
-        const onDeviceChange = config.onDeviceChange ?? "replace";
-        const camera = client.devices.camera;
-        const stream = camera.broadcast?.stream;
-
-        if (isBroadcastedDeviceChanged("userMedia", client, pending, event.mediaDeviceType)) {
-          if (!stream && config.broadcastOnDeviceStart) {
-            pending = true;
-
-            await camera
-              .addTrack(config.defaultTrackMetadata, config.defaultSimulcastConfig, config.defaultMaxBandwidth)
-              .finally(() => {
-                pending = false;
-              });
-          } else if (stream && onDeviceChange === "replace") {
-            pending = true;
-
-            await camera.replaceTrack().finally(() => {
-              pending = false;
-            });
-          } else if (stream && onDeviceChange === "remove") {
-            pending = true;
-
-            await camera.removeTrack().finally(() => {
-              pending = false;
-            });
-          }
-        }
-      };
-
-      const managerInitialized: ClientEvents<PeerMetadata, TrackMetadata>["managerInitialized"] = async (
-        event,
-        client,
-      ) => {
-        if (event.video?.media?.stream) {
-          await broadcastOnCameraStart(event, client);
-        }
-      };
-
-      const devicesReady: ClientEvents<PeerMetadata, TrackMetadata>["devicesReady"] = async (event, client) => {
-        if (event.video.restarted && event.video?.media?.stream) {
-          await broadcastOnCameraStart(event, client);
-        }
-      };
-
-      const deviceReady: ClientEvents<PeerMetadata, TrackMetadata>["deviceReady"] = async (event, client) => {
-        if (event.trackType === "video") {
-          await broadcastOnCameraStart(event, client);
-        }
-      };
-
-      state.client.on("managerInitialized", managerInitialized);
-      state.client.on("devicesReady", devicesReady);
-      state.client.on("deviceReady", deviceReady);
-
-      return () => {
-        state.client.removeListener("managerInitialized", managerInitialized);
-        state.client.removeListener("devicesReady", devicesReady);
-        state.client.removeListener("deviceReady", deviceReady);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      const removeOnCameraStopped: ClientEvents<PeerMetadata, TrackMetadata>["deviceStopped"] = async (
-        event,
-        client,
-      ) => {
-        const camera = client.devices.camera;
-        const stream = camera.broadcast?.stream;
-        const onDeviceStop = configRef.current.camera.onDeviceStop ?? "mute";
-
-        if (isBroadcastedTrackStopped("userMedia", "video", client.status, event, stream)) {
-          if (onDeviceStop === "mute") {
-            await camera.muteTrack();
-          } else {
-            await camera.removeTrack();
-          }
-        }
-      };
-
-      state.client.on("deviceStopped", removeOnCameraStopped);
-
-      return () => {
-        state.client.removeListener("deviceStopped", removeOnCameraStopped);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      const broadcastCameraOnConnect: ClientEvents<PeerMetadata, TrackMetadata>["joined"] = async (_, client) => {
-        const camera = client.devices.camera;
-        const stream = camera.stream;
-        const config = configRef.current.camera;
-
-        if (stream && config.broadcastOnConnect) {
-          await camera.addTrack(config.defaultTrackMetadata, config.defaultSimulcastConfig, config.defaultMaxBandwidth);
-        }
-      };
-
-      state.client.on("joined", broadcastCameraOnConnect);
-
-      return () => {
-        state.client.removeListener("joined", broadcastCameraOnConnect);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      let pending = false;
-
-      const broadcastOnMicrophoneStart = async (
-        event: { mediaDeviceType: MediaDeviceType },
-        client: ClientApi<PeerMetadata, TrackMetadata>,
-      ) => {
-        const microphone = client.devices.microphone;
-        const stream = microphone.broadcast?.stream;
-        const config = configRef.current.microphone;
-        const onDeviceChange = config.onDeviceChange ?? "replace";
-
-        if (isBroadcastedDeviceChanged("userMedia", client, pending, event.mediaDeviceType)) {
-          if (!stream && config.broadcastOnDeviceStart) {
-            pending = true;
-
-            await microphone.addTrack(config.defaultTrackMetadata, config.defaultMaxBandwidth).finally(() => {
-              pending = false;
-            });
-          } else if (stream && onDeviceChange === "replace") {
-            pending = true;
-
-            await microphone.replaceTrack().finally(() => {
-              pending = false;
-            });
-          } else if (stream && onDeviceChange === "remove") {
-            pending = true;
-
-            await microphone.removeTrack().finally(() => {
-              pending = false;
-            });
-          }
-        }
-      };
-
-      const managerInitialized: ClientEvents<PeerMetadata, TrackMetadata>["managerInitialized"] = async (
-        event,
-        client,
-      ) => {
-        if (event.audio?.media?.stream) {
-          await broadcastOnMicrophoneStart(event, client);
-        }
-      };
-
-      const devicesReady: ClientEvents<PeerMetadata, TrackMetadata>["devicesReady"] = async (event, client) => {
-        if (event.audio.restarted && event.audio?.media?.stream) {
-          await broadcastOnMicrophoneStart(event, client);
-        }
-      };
-
-      const deviceReady: ClientEvents<PeerMetadata, TrackMetadata>["deviceReady"] = async (event, client) => {
-        if (event.trackType === "audio") {
-          await broadcastOnMicrophoneStart(event, client);
-        }
-      };
-
-      state.client.on("managerInitialized", managerInitialized);
-      state.client.on("deviceReady", deviceReady);
-      state.client.on("devicesReady", devicesReady);
-
-      return () => {
-        state.client.removeListener("managerInitialized", managerInitialized);
-        state.client.removeListener("deviceReady", deviceReady);
-        state.client.removeListener("devicesReady", devicesReady);
-      };
-    }, [state.client]);
-
-    const isBroadcastedTrackStopped = (
-      expectedMediaDeviceType: MediaDeviceType,
-      expectedTrackType: TrackType,
-      status: PeerStatus,
-      event: Parameters<ClientEvents<PeerMetadata, TrackMetadata>["deviceStopped"]>[0],
-      stream: MediaStream | undefined | null,
-    ) =>
-      status === "joined" &&
-      event.mediaDeviceType === expectedMediaDeviceType &&
-      event.trackType === expectedTrackType &&
-      stream;
-
-    useEffect(() => {
-      const onMicrophoneStopped: ClientEvents<PeerMetadata, TrackMetadata>["deviceStopped"] = async (event, client) => {
-        const microphone = client.devices.microphone;
-        const stream = microphone.broadcast?.stream;
-        const onDeviceStop = configRef.current.microphone.onDeviceStop ?? "mute";
-
-        if (isBroadcastedTrackStopped("userMedia", "audio", client.status, event, stream)) {
-          if (onDeviceStop === "mute") {
-            await microphone.muteTrack();
-          } else {
-            await microphone.removeTrack();
-          }
-        }
-      };
-
-      state.client.on("deviceStopped", onMicrophoneStopped);
-
-      return () => {
-        state.client.removeListener("deviceStopped", onMicrophoneStopped);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      const broadcastMicrophoneOnConnect: ClientEvents<PeerMetadata, TrackMetadata>["joined"] = async (_, client) => {
-        const config = configRef.current.microphone;
-        const microphone = client.devices.microphone;
-
-        if (microphone.stream && config.broadcastOnConnect) {
-          await microphone.addTrack(config.defaultTrackMetadata, config.defaultMaxBandwidth);
-        }
-      };
-
-      state.client.on("joined", broadcastMicrophoneOnConnect);
-
-      return () => {
-        state.client.removeListener("joined", broadcastMicrophoneOnConnect);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      let pending = false;
-
-      const broadcastOnScreenShareStart: ClientEvents<PeerMetadata, TrackMetadata>["deviceReady"] = async (
-        event: { mediaDeviceType: MediaDeviceType },
-        client,
-      ) => {
-        const screenShare = client.devices.screenShare;
-        const stream = screenShare.broadcast?.stream;
-        const { broadcastOnDeviceStart, defaultTrackMetadata, defaultMaxBandwidth } = configRef.current.screenShare;
-
-        if (
-          isBroadcastedDeviceChanged("displayMedia", client, pending, event.mediaDeviceType) &&
-          !stream &&
-          broadcastOnDeviceStart
-        ) {
-          pending = true;
-
-          await screenShare.addTrack(defaultTrackMetadata, defaultMaxBandwidth).finally(() => {
-            pending = false;
-          });
-        }
-      };
-
-      state.client.on("deviceReady", broadcastOnScreenShareStart);
-
-      return () => {
-        state.client.removeListener("deviceReady", broadcastOnScreenShareStart);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      const onScreenShareStop: ClientEvents<PeerMetadata, TrackMetadata>["deviceStopped"] = async (event, client) => {
-        const stream = client.devices.screenShare.broadcast?.stream;
-        if (isBroadcastedTrackStopped("displayMedia", "video", client.status, event, stream)) {
-          await client.devices.screenShare.removeTrack();
-        }
-      };
-
-      state.client.on("deviceStopped", onScreenShareStop);
-
-      return () => {
-        state.client.removeListener("deviceStopped", onScreenShareStop);
-      };
-    }, [state.client]);
-
-    useEffect(() => {
-      const broadcastScreenShareOnConnect: ClientEvents<PeerMetadata, TrackMetadata>["joined"] = async (_, client) => {
-        if (client.devices.screenShare.stream && configRef.current.screenShare.broadcastOnConnect) {
-          await client.devices.screenShare.addTrack(
-            configRef.current.screenShare.defaultTrackMetadata,
-            configRef.current.screenShare.defaultMaxBandwidth,
-          );
-        }
-      };
-
-      state.client.on("joined", broadcastScreenShareOnConnect);
-
-      return () => {
-        state.client.removeListener("joined", broadcastScreenShareOnConnect);
-      };
-    }, [state.client]);
-
-    return useMemo(
-      () => ({
-        init: () =>
-          state.devices.init({
-            audioTrackConstraints: configRef.current?.microphone?.trackConstraints,
-            videoTrackConstraints: configRef.current?.camera?.trackConstraints,
-          }),
-      }),
-      [state.devices],
-    );
-  };
-
   const useScreenShare = (): ScreenShareAPI<TrackMetadata> => {
     const { state } = useFishjamContext();
     return state.devices.screenShare;
@@ -630,7 +266,7 @@ export const create = <PeerMetadata, TrackMetadata>(
     useDisconnect,
     useStatus,
     useTracks,
-    useSetupMedia,
+    useSetupMedia: createUseSetupMediaHook(useFishjamContext),
     useCamera,
     useMicrophone,
     useScreenShare,
